# docs/OPENAPI_SPEC.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Healthcare Services Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/prescriptions/{prescription_id}/verify": {
      "get": {
        "tags": [
          "Chemists"
        ],
        "summary": "Verify Prescription Signature Endpoint",
        "description": "Public / Chemist endpoint to cryptographically verify digital signature integrity.",
        "operationId": "verify_prescription_signature_endpoint_api_v1_prescriptions__prescription_id__verify_get",
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Prescription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionVerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/chemists/register": {
      "post": {
        "tags": [
          "Chemists"
        ],
        "summary": "Register Chemist",
        "description": "Register a pharmacy with Drug License (DL) details.",
        "operationId": "register_chemist_api_v1_chemists_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChemistRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChemistResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chemists": {
      "get": {
        "tags": [
          "Chemists"
        ],
        "summary": "List Chemists",
        "description": "List directory of pharmacies with optional locality/city/clinic filters.",
        "operationId": "list_chemists_api_v1_chemists_get",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "locality",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Locality"
            }
          },
          {
            "name": "clinic_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Clinic Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChemistResponse"
                  },
                  "title": "Response List Chemists Api V1 Chemists Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chemist/prescriptions": {
      "get": {
        "tags": [
          "Chemists"
        ],
        "summary": "List Chemist Prescriptions",
        "description": "View digital prescriptions routed to the logged-in chemist.",
        "operationId": "list_chemist_prescriptions_api_v1_chemist_prescriptions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dispense_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dispense Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrescriptionResponse"
                  },
                  "title": "Response List Chemist Prescriptions Api V1 Chemist Prescriptions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chemist/prescriptions/{prescription_id}/dispense": {
      "post": {
        "tags": [
          "Chemists"
        ],
        "summary": "Dispense Prescription",
        "description": "Chemist marks prescription as fulfilled / dispensed.",
        "operationId": "dispense_prescription_api_v1_chemist_prescriptions__prescription_id__dispense_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prescription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Prescription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispensePrescriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChemistRegisterRequest": {
        "properties": {
          "pharmacy_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Pharmacy Name"
          },
          "license_number": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "License Number"
          },
          "clinic_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinic Id"
          },
          "address": {
            "type": "string",
            "minLength": 5,
            "title": "Address"
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "City"
          },
          "locality": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Locality"
          },
          "pincode": {
            "type": "string",
            "maxLength": 10,
            "minLength": 5,
            "title": "Pincode"
          },
          "contact_number": {
            "type": "string",
            "maxLength": 20,
            "minLength": 8,
            "title": "Contact Number"
          }
        },
        "type": "object",
        "required": [
          "pharmacy_name",
          "license_number",
          "address",
          "city",
          "locality",
          "pincode",
          "contact_number"
        ],
        "title": "ChemistRegisterRequest"
      },
      "ChemistResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "pharmacy_name": {
            "type": "string",
            "title": "Pharmacy Name"
          },
          "license_number": {
            "type": "string",
            "title": "License Number"
          },
          "clinic_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinic Id"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "locality": {
            "type": "string",
            "title": "Locality"
          },
          "pincode": {
            "type": "string",
            "title": "Pincode"
          },
          "contact_number": {
            "type": "string",
            "title": "Contact Number"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "pharmacy_name",
          "license_number",
          "address",
          "city",
          "locality",
          "pincode",
          "contact_number",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "ChemistResponse"
      },
...
      "DispensePrescriptionResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "prescription_id": {
            "type": "string",
            "format": "uuid",
            "title": "Prescription Id"
          },
          "dispense_status": {
            "type": "string",
            "title": "Dispense Status"
          },
          "dispensed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Dispensed At"
          },
          "dispensed_by_chemist_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dispensed By Chemist Id"
          }
        },
        "type": "object",
        "required": [
          "message",
          "prescription_id",
          "dispense_status",
          "dispensed_at",
          "dispensed_by_chemist_id"
        ],
        "title": "DispensePrescriptionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrescriptionItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "drug_name": {
            "type": "string",
            "title": "Drug Name"
          },
          "dosage": {
            "type": "string",
            "title": "Dosage"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency"
          },
          "duration_days": {
            "type": "integer",
            "title": "Duration Days"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "drug_name",
          "dosage",
          "frequency",
          "duration_days"
        ],
        "title": "PrescriptionItemResponse"
      },
      "PrescriptionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "appointment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Appointment Id"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "diagnosis": {
            "type": "string",
            "title": "Diagnosis"
          },
          "clinical_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinical Notes"
          },
          "pdf_s3_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf S3 Key"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At"
          },
          "chemist_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chemist Id"
          },
          "digital_signature": {
            "type": "string",
            "title": "Digital Signature"
          },
          "digital_signature_timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Digital Signature Timestamp"
          },
          "dispense_status": {
            "type": "string",
            "title": "Dispense Status",
            "default": "pending"
          },
          "dispensed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispensed At"
          },
          "dispensed_by_chemist_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispensed By Chemist Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PrescriptionItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "appointment_id",
          "doctor_id",
          "patient_id",
          "diagnosis",
          "issued_at",
          "digital_signature",
          "digital_signature_timestamp",
          "items",
          "created_at",
          "updated_at"
        ],
        "title": "PrescriptionResponse"
      },
      "PrescriptionVerificationResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "prescription_id": {
            "type": "string",
            "format": "uuid",
            "title": "Prescription Id"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "appointment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Appointment Id"
          },
          "doctor_reg_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor Reg Number"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At"
          },
          "digital_signature": {
            "type": "string",
            "title": "Digital Signature"
          },
          "dispense_status": {
            "type": "string",
            "title": "Dispense Status"
          },
          "dispensed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispensed At"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "valid",
          "prescription_id",
          "doctor_id",
          "patient_id",
          "appointment_id",
          "issued_at",
          "digital_signature",
          "dispense_status",
          "message"
        ],
        "title": "PrescriptionVerificationResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}